LayoutTests/fast/canvas/canvas-blending-gradient-over-color.html

<!DOCTYPE HTML>
<html>
<body>
    <script src="../js/resources/js-test-pre.js"></script>
    <script type="text/javascript" src="canvas-blending-helpers.js"></script>
    <script type="text/javascript">

        description("Series of tests to ensure correct results on applying different blend modes when drawing a gradient on top of a color.");

        function checkBlendModeResult(i, context, sigma) {
            var expectedColor = blendColors([129 / 255, 1, 129 / 255, 1], [1, 129 / 255, 129 / 255, 1], i);

            var actualColor = context.getImageData(0, 0, 1, 1).data;
            shouldBeCloseTo(actualColor[0] + "", expectedColor[0], sigma);
            shouldBeCloseTo(actualColor[1] + "", expectedColor[1], sigma);
            shouldBeCloseTo(actualColor[2] + "", expectedColor[2], sigma);
            shouldBeCloseTo(actualColor[3] + "", expectedColor[3], sigma);
        }

        function runTest() {
            var canvas = document.createElement("canvas");
            var sigma = 5;
            canvas.width = 10;
            canvas.height = 10;
            var context = canvas.getContext("2d");

            for (var i = 0; i < blendModes.length; ++i) {
                debug("Testing blend mode " + blendModes[i]);

                context.clearRect(0, 0, 10, 10);
                context.save();
                drawBackdropColorInContext(context);
                if (blendModes[i] != "soft-light")
                    context.globalCompositeOperation = blendModes[i];
                drawSourceColorGradientInContext(context);
                checkBlendModeResult(i, context, sigma);
                context.restore();
                debug('');
            }
        }

        runTest();
    </script>
    <script src="../js/resources/js-test-post.js"></script>
</body>
</html>
